Partial implementation of support for resource ID parsing. Partial implementation of support for Swagger-based API documentation.

+++ src/main/java/com/whizzosoftware/hobson/rest/v1/resource/device/DeviceResource.java
@@ -1,17 +1,23 @@
-/*******************************************************************************
+/*
+ *******************************************************************************
  * Copyright (c) 2014 Whizzo Software, LLC.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
- *******************************************************************************/
+ *******************************************************************************
+*/
 package com.whizzosoftware.hobson.rest.v1.resource.device;
 
+import com.whizzosoftware.hobson.api.HobsonAuthorizationException;
 import com.whizzosoftware.hobson.api.device.DeviceContext;
 import com.whizzosoftware.hobson.api.device.DeviceManager;
+import com.whizzosoftware.hobson.api.user.HobsonRole;
 import com.whizzosoftware.hobson.dto.ExpansionFields;
+import com.whizzosoftware.hobson.dto.context.DTOBuildContext;
 import com.whizzosoftware.hobson.dto.context.DTOBuildContextFactory;
 import com.whizzosoftware.hobson.dto.device.HobsonDeviceDTO;
+import com.whizzosoftware.hobson.json.JSONAttributes;
 import com.whizzosoftware.hobson.rest.HobsonAuthorizer;
 import com.whizzosoftware.hobson.rest.HobsonRestContext;
 import com.whizzosoftware.hobson.rest.v1.util.MediaTypeHelper;
@@ -30,55 +36,27 @@ import javax.inject.Inject;
  * @author Dan Noguerol
  */
 public class DeviceResource extends SelfInjectingServerResource {
-    public static final String PATH = "/hubs/{hubId}/plugins/{pluginId}/devices/{deviceId}";
+    public static final String PATH = "/hubs/{hubId}/plugins/local/{pluginId}/devices/{deviceId}";
 
     @Inject
     DeviceManager deviceManager;
     @Inject
     DTOBuildContextFactory dtoBuildContextFactory;
 
-    /**
-     * @api {get} /api/v1/users/:userId/hubs/:hubId/plugins/:pluginId/devices/:deviceId Get device details
-     * @apiVersion 0.1.3
-     * @apiName GetDeviceDetails
-     * @apiDescription Retrieves the details of a specific device.
-     * @apiGroup Devices
-     * @apiParam (Query Parameters) {String} expand A comma-separated list of attributes to expand (supported values are "configuration", "configurationClass", "preferredVariable", "variables").
-     * @apiSuccess {String} name The device name.
-     * @apiSuccess {String} type The device type.
-     * @apiSuccess {Object} configuration The current configuration values for the device.
-     * @apiSuccess {Object} configurationClass The device's configuration class.
-     * @apiSuccess {Object} preferredVariable The device's "preferred variable" if it has one.
-     * @apiSuccess {Object} variables The variables the device has published.
-     * @apiSuccessExample {json} Success Response:
-     * {
-     *   "name": "RadioRa Zone 1",
-     *   "type": "LIGHTBULB",
-     *   "configuration": {
-     *     "@id": "/api/plugins/com.whizzosoftware.hobson.hub.hobson-hub-radiora/devices/1/configuration"
-     *   },
-     *   "configurationClass": {
-     *     "@id": "/api/plugins/com.whizzosoftware.hobson.hub.hobson-hub-radiora/devices/1/configurationClass"
-     *   },
-     *   "preferredVariable": {
-     *     "@id": "/api/plugins/com.whizzosoftware.hobson.hub.hobson-hub-radiora/devices/1/variables/on"
-     *   },
-     *   "variables": {
-     *     "@id": "/api/plugins/com.whizzosoftware.hobson.hub.hobson-hub-radiora/devices/1/variables"
-     *   }
-     * }
-     */
     @Override
     protected Representation get() throws ResourceException {
         HobsonRestContext ctx = (HobsonRestContext)getRequest().getAttributes().get(HobsonAuthorizer.HUB_CONTEXT);
         ExpansionFields expansions = new ExpansionFields(getQueryValue("expand"));
+        DTOBuildContext dbctx = dtoBuildContextFactory.createContext(ctx.getApiRoot(), expansions);
 
         DeviceContext dctx = DeviceContext.create(ctx.getHubContext(), getAttribute("pluginId"), getAttribute("deviceId"));
         HobsonDeviceDTO dto = new HobsonDeviceDTO.Builder(
-            dtoBuildContextFactory.createContext(ctx.getApiRoot(), expansions),
+            dbctx,
             dctx,
             true
         ).build();
+
+        dto.addContext(JSONAttributes.AIDT, dbctx.getIdTemplateMap());
 
         JsonRepresentation jr = new JsonRepresentation(dto.toJSON());
         jr.setMediaType(MediaTypeHelper.createMediaType(getRequest(), dto));
@@ -87,6 +65,10 @@ public class DeviceResource extends SelfInjectingServerResource {
 
     @Override
     protected Representation delete() throws ResourceException {
+        if (!isInRole(HobsonRole.administrator.name()) && !isInRole(HobsonRole.userWrite.name())) {
+            throw new HobsonAuthorizationException("Forbidden");
+        }
+
         HobsonRestContext ctx = (HobsonRestContext)getRequest().getAttributes().get(HobsonAuthorizer.HUB_CONTEXT);
         deviceManager.deleteDevice(DeviceContext.create(ctx.getHubContext(), getAttribute("pluginId"), getAttribute("deviceId")));
         getResponse().setStatus(Status.SUCCESS_ACCEPTED);
